task-management-app/backend/task_service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
